Refactor editor page into modular components and hooks
The editor/[id] page had grown large and unwieldy, making it difficult to maintain and scale. This refactoring breaks down the monolithic component into smaller, logically separated pieces to improve readability, maintainability, and testability.

**Key Changes:**
- Extract types to `types/index.ts` for centralized type definitions
- Create utility modules:
  - `utils/fileTreeHelpers.ts` - File tree manipulation functions
  - `utils/colorHelpers.ts` - Color utility functions
  - `utils/monacoSetup.ts` - Monaco editor configuration
- Implement custom hooks:
  - `hooks/useUser.ts` - User authentication state
  - `hooks/useSocket.ts` - Socket.io communication
  - `hooks/useFileTree.ts` - File tree state management
  - `hooks/useYjs.ts` - Yjs and Monaco integration
- Split UI into focused components:
  - `components/FileTree.tsx` - File tree navigation
  - `components/ChatPanel.tsx` - Real-time chat interface
  - `components/PresenceList.tsx` - User presence display
- Simplify main `page.tsx` to act as coordinator using the new hooks and components

**Improvements:**
- Better separation of concerns
- Enhanced reusability of hooks and components
- Improved code organization and readability
- Easier testing and debugging
- More maintainable codebase structure

This refactoring maintains all existing functionality while making the code much more modular and easier to work with.
